Solution to assignment
makeCacheMatrix initialises the solution to NULL, creates the getter and
setter functions for the matrix and its solution and returns a list of
the methods.

cacheSolve is a function which returns the inverse of a matrix
## First it checks for a cached solution and if this exists, it returns
it. Otherwise it calculates the solution, stores the result in the cache
and returns the solution.